requirements.txt
flask==3.0.0
pdfplumber==0.11.0
python-docx==1.1.0
gunicorn==21.2.0